feat: Refactor Hover spec
