--- spec/ruby_lsp/support/hovers/data/sample_spec.rb
@@ -1,2 +1,0 @@
-describe SampleSpec do
-end
